+++ src/main/java/spring/app/controller/restController/AdminRestController.java
@@ -6,14 +6,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import spring.app.dto.CompanyDto;
 import spring.app.dto.UserDto;
-import spring.app.model.Company;
-import spring.app.model.OrgType;
-import spring.app.model.Role;
-import spring.app.model.User;
-import spring.app.service.abstraction.CompanyService;
-import spring.app.service.abstraction.GenreService;
-import spring.app.service.abstraction.RoleService;
-import spring.app.service.abstraction.UserService;
+import spring.app.model.*;
+import spring.app.service.abstraction.*;
 
 import java.time.LocalTime;
 import java.util.HashSet;
@@ -30,13 +24,15 @@ public class AdminRestController {
     private final UserService userService;
     private final CompanyService companyService;
     private final GenreService genreService;
+    private final AuthorService authorService;
 
     @Autowired
-    public AdminRestController(RoleService roleService, UserService userService, CompanyService companyService, GenreService genreService) {
+    public AdminRestController(RoleService roleService, UserService userService, CompanyService companyService, GenreService genreService, AuthorService authorService) {
         this.roleService = roleService;
         this.userService = userService;
         this.companyService = companyService;
         this.genreService = genreService;
+        this.authorService = authorService;
     }
 
     @GetMapping(value = "/all_users")
@@ -53,11 +49,27 @@ public class AdminRestController {
         return list;
     }
 
+    @GetMapping(value = "/all_authors")
+    public List<Author> getAllAuthor(){
+        List<Author> list = authorService.getAllAuthor();
+        return list;
+    }
+
     @PostMapping(value = "/add_user")
     public void addUser(@RequestBody UserDto userDto) {
         User user = new User(userDto.getEmail(), userDto.getLogin(), userDto.getPassword(), true);
         user.setRoles(getRoles(userDto.getRoles()));
         userService.addUser(user);
+    }
+
+    @PostMapping(value = "/add_author")
+    public void addAuthor(@RequestBody String name){
+        name = name.replaceAll("[^A-Za-zА-Яа-я0-9 ]", "");
+        if (genreService.getByName(name) == null) {
+            Author author = new Author();
+            author.setName(name);
+            authorService.addAuthor(author);
+        }
     }
 
     @PutMapping(value = "/update_user")
@@ -67,9 +79,21 @@ public class AdminRestController {
         userService.updateUser(user);
     }
 
+    @PutMapping(value = "/update_author")
+    public void updateAuthor(@RequestBody Author newAuthor){
+        Author author = authorService.getById(newAuthor.getId());
+        author.setName(newAuthor.getName());
+        authorService.updateAuthor(author);
+    }
+
     @DeleteMapping(value = "/delete_user")
     public void deleteUser(@RequestBody Long id) {
         userService.deleteUserById(id);
+    }
+
+    @DeleteMapping(value = "/delete_author")
+    public void deleteAuthor(@RequestBody Long id){
+        authorService.deleteAuthorById(id);
     }
 
     @GetMapping(value = "/company/{id}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
