+++ src/main/resources/templates/admin/performers.html
@@ -82,6 +82,8 @@
                                                 <tr>
                                                     <th>ID</th>
                                                     <th>Имя</th>
+                                                    <th>Изменить</th>
+                                                    <th>Удалить</th>
                                                 </tr>
                                             </thead>
 
@@ -158,7 +160,6 @@
                         </div>
                     </div>
                 </div>
-<!--                ///////-->
             </div>
         </div>
     </div>
